Refractor: moved my physics logic into Object3D class
Decided to clean up my code and add more properties to my Object3D class in order to handle physics simulation. Added a isMoving property so that the physics is only applied to objects that I want to move. I also implemented setters and getters for them. Included a tick() function as shown in lecture.